# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  api.BuildEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.BuildResponse'
    type: object
...
  api.BuildLogResponse:
    properties:
      message:
        type: string
      step_name:
        type: string
      timestamp:
        type: string
    type: object
  api.BuildLogsEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.BuildLogsResponse'
    type: object
  api.BuildLogsResponse:
    properties:
      build_id:
        type: string
      logs:
        items:
          $ref: '#/definitions/api.BuildLogResponse'
        type: array
    type: object
  api.BuildResponse:
    properties:
      actor:
        type: string
      attempt_number:
        type: integer
      build_number:
        type: integer
      created_at:
        type: string
      current_step_index:
        type: integer
      delivery_id:
        type: string
      error_message:
        type: string
      event_type:
        type: string
      finished_at:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/api.ImageExecutionResponse'
      job_id:
        type: string
      pipeline_config_yaml:
        type: string
      pipeline_name:
        type: string
      pipeline_path:
        type: string
      pipeline_source:
        type: string
      priority:
        type: integer
      project_id:
        type: string
      project_name:
        type: string
      project_slug:
        type: string
      queued_at:
        type: string
      ref_type:
        type: string
      repository_name:
        type: string
      repository_owner:
        type: string
      repository_url:
        type: string
      rerun_from_step_index:
        type: integer
      rerun_of_build_id:
        type: string
      scm_provider:
        type: string
      source:
        $ref: '#/definitions/api.BuildSourceResponse'
      source_author_email:
        type: string
      source_author_name:
        type: string
      source_commit_sha:
        type: string
      source_committer_email:
        type: string
      source_committer_name:
        type: string
      source_ref:
        type: string
      source_sha:
        type: string
      started_at:
        type: string
      status:
        type: string
      trigger_commit_sha:
        type: string
      trigger_kind:
        type: string
      trigger_ref:
        type: string
      trigger_type:
        type: string
      triggered_by:
        type: string
    type: object
...
  api.BuildSourceResponse:
    properties:
      ref:
        type: string
      repository_url:
        type: string
      source_commit_sha:
        type: string
    type: object
  api.BuildStepResponse:
    properties:
      build_id:
        type: string
      command:
        type: string
      error_message:
        type: string
      exit_code:
        type: integer
      finished_at:
        type: string
      group_name:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/api.ImageExecutionResponse'
      job:
        $ref: '#/definitions/api.ExecutionJobResponse'
      name:
        type: string
      started_at:
        type: string
      status:
        type: string
      stderr:
        type: string
      stdout:
        type: string
      step_index:
        type: integer
      worker_id:
        type: string
    type: object
  api.BuildStepsEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.BuildStepsResponse'
    type: object
  api.BuildStepsResponse:
    properties:
      build_id:
        type: string
      steps:
        items:
          $ref: '#/definitions/api.BuildStepResponse'
        type: array
    type: object
...
  api.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.ErrorBody'
    type: object
  api.ExecutionJobOutputResponse:
    properties:
      build_id:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      declared_path:
        type: string
      destination_uri:
        type: string
      digest:
        type: string
      id:
        type: string
      job_id:
        type: string
      kind:
        type: string
      name:
        type: string
      size_bytes:
        type: integer
      status:
        type: string
    type: object
  api.ExecutionJobResponse:
    properties:
      attempt_number:
        type: integer
      build_id:
        type: string
      command:
        items:
          type: string
        type: array
      command_preview:
        type: string
      context_dir:
        type: string
      created_at:
        type: string
      environment:
        additionalProperties:
          type: string
        type: object
      error_message:
        type: string
      finished_at:
        type: string
      id:
        type: string
      image:
        type: string
      lineage_root_job_id:
        type: string
      name:
        type: string
      outputs:
        items:
          $ref: '#/definitions/api.ExecutionJobOutputResponse'
        type: array
      pipeline_file_path:
        type: string
      retry_of_job_id:
        type: string
      source_commit_sha:
        type: string
      source_ref_name:
        type: string
      source_repo_url:
        type: string
      spec_digest:
        type: string
      spec_version:
        type: integer
      started_at:
        type: string
      status:
        type: string
      step_id:
        type: string
      step_index:
        type: integer
      timeout_seconds:
        type: integer
      working_dir:
        type: string
    type: object
  api.ImageExecutionResponse:
    properties:
      managed_image_id:
        type: string
      managed_image_version_id:
        type: string
      requested_ref:
        type: string
      resolved_ref:
        type: string
      source_kind:
        type: string
      version_tags:
        items:
          $ref: '#/definitions/api.VersionTagResponse'
        type: array
    type: object
...
  api.QueueBuildRequest:
    properties:
      steps:
        items:
          $ref: '#/definitions/api.QueueBuildStepInput'
        type: array
      template:
        type: string
    type: object
  api.QueueBuildStepInput:
    properties:
      command:
        type: string
      name:
        type: string
    type: object
...
  api.RetryJobEnvelope:
    properties:
      data:
        $ref: '#/definitions/api.RetryJobResponse'
    type: object
  api.RetryJobResponse:
    properties:
      build:
        $ref: '#/definitions/api.BuildResponse'
      job:
        $ref: '#/definitions/api.ExecutionJobResponse'
    type: object
...
  api.VersionTagResponse:
    properties:
      artifact_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      job_id:
        type: string
      kind:
        type: string
      managed_image_version_id:
        type: string
      target_type:
        type: string
      version:
        type: string
    type: object
info:
  contact: {}
  description: HTTP API for Coyote CI control-plane workflows.
  title: Coyote CI API
  version: "0.1"
paths:
...
  /builds/{buildID}/cancel:
    post:
      description: Marks a queued or running build as canceled and terminalizes cancellable
        steps/jobs.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Cancel build
      tags:
      - builds
  /builds/{buildID}/complete:
    post:
      description: Transitions build status from running to success.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Complete build
      tags:
      - builds
  /builds/{buildID}/fail:
    post:
      description: Transitions build status from running to failed.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Fail build
      tags:
      - builds
  /builds/{buildID}/logs:
    get:
      description: Returns log lines for a build.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildLogsEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get build logs
      tags:
      - builds
  /builds/{buildID}/queue:
    post:
      consumes:
      - application/json
      description: Transitions build status from pending to queued.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      - description: Queue build request
        in: body
        name: request
        schema:
          $ref: '#/definitions/api.QueueBuildRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Queue build
      tags:
      - builds
  /builds/{buildID}/rerun:
    post:
      description: Creates a new queued build using the source, trigger, job, and
        step context from an existing build.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Rerun build
      tags:
      - builds
  /builds/{buildID}/start:
    post:
      description: Transitions build status from queued to running.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Start build
      tags:
      - builds
  /builds/{buildID}/steps:
    get:
      description: Returns steps for a build.
      parameters:
      - description: Build ID
        in: path
        name: buildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BuildStepsEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get build steps
      tags:
      - builds
...
  /builds/jobs/{jobID}/retry:
    post:
      description: Creates a new build attempt containing a retry of the failed execution
        job.
      parameters:
      - description: Execution Job ID
        in: path
        name: jobID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RetryJobEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Retry failed execution job
      tags:
      - builds
...
schemes:
- http
swagger: "2.0"